+++ particles-kool/src/commonMain/kotlin/me/dvyy/particles/CommonLauncher.kt
@@ -1,13 +1,12 @@
 package me.dvyy.particles
 
 import de.fabmax.kool.KoolContext
+import de.fabmax.kool.modules.ui2.*
+import de.fabmax.kool.util.Color
 import de.fabmax.kool.util.debugOverlay
 import me.dvyy.particles.compute.forces.Force
-import me.dvyy.particles.compute.forces.PairwiseForce
-import me.dvyy.particles.config.AppSettings
-import me.dvyy.particles.config.ConfigRepository
-import org.koin.core.module.dsl.singleOf
-import org.koin.dsl.module
+import me.dvyy.particles.ui.AppSizes
+import me.dvyy.particles.ui.helpers.TRANSPARENT
 
 fun launchApp(ctx: KoolContext, forces: List<Force>) {
     val baseModule = persistentModule(ctx)
@@ -15,6 +14,23 @@ fun launchApp(ctx: KoolContext, forces: List<Force>) {
     val manager = SceneManager(ctx, baseModule, forces)
 
     ctx.scenes += debugOverlay()
+    ctx.scenes += UiScene("version-info") {
+        addPanelSurface(
+            sizes = Sizes.small,
+            colors = Colors.darkColors(primary = Color("b2ff00"), background = Color.TRANSPARENT),
+            width = FitContent,
+            height = FitContent
+        ) {
+            modifier.align(AlignmentX.End, AlignmentY.Bottom)
+                .padding(vertical = 1.dp)
+                .padding(end = AppSizes.sidebarSize + 1.dp)
+            Text("v${BuildKonfig.version}") {
+                modifier
+                    .align(AlignmentX.End, AlignmentY.Bottom)
+                    .textColor(Color.WHITE.withAlpha(0.6f))
+            }
+        }
+    }
     manager.load()
 }